Update clean neutrophil and run DGEA
The previous 240202_lav_neut_n8_2000_log_cfam_clean_S3.rds did not
exist for some reason, so re-created it from line 41 of this commit.
Generated figures for the clean object, then moved onto DGE analysis.
This revealed notable changes, but with opposing (or otherwise independent)
changes when comparing d14 to d0 and d90 to d0. This is conistent with
earlier analysis and provides strong evidence that d90 and d14 samples
should NOT be collapsed into one post-treatment group.

TO DO: look closer to ensure something is not being flipped, as this is an
unexpected finding.