Add extraction and plotting of inferCNV output using HMM
Previously, the heatmap data was extracted and used to split the cells
into quartiles using CNV burden. This apporach was not ideal and apprently
there is a built in function to get at relavtive CNV burden. So, this commit
takes advantage of the approach and also clean up the old approach (Obsolete)
portion of the script - will be deleted in time.