Add filtering of low quality cells
Several cells that appeared to be myeloid were still in the epithelial subset
so this commit remove all potential contaminates. This call was based on CD4,
PTPRC, and AIF1 expression.

In removing cells there were inadequate cell numbers for "nasal_pt6", so
somthing will have to be done to bring over the annotation of the removed cells.

There a likely too few cells to do much with analysis of this subpopulation.